+++ init.go
@@ -29,7 +29,7 @@ func init() {
 	framework.PortRangeFrom, _ = strconv.Atoi(util.Getenv("PORTRANGE_FROM", "31000"))
 	framework.PortRangeTo, _ = strconv.Atoi(util.Getenv("PORTRANGE_TO", "32000"))
 	config.Principal = util.Getenv("MESOS_PRINCIPAL", "")
-	config.CPU, _ = strconv.ParseFloat(util.Getenv("DEFAULT_CPU", "0.001"), 64)
+	config.CPU, _ = strconv.ParseFloat(util.Getenv("DEFAULT_CPU", "0.1"), 64)
 	config.Memory, _ = strconv.ParseFloat(util.Getenv("DEFAULT_MEMORY", "50"), 64)
 	config.Disk, _ = strconv.ParseFloat(util.Getenv("DEFAULT_DISK", "1000"), 64)
 	config.DefaultVolumeDriver = util.Getenv("DEFAULT_VOLUME_DRIVER", "local")