+++ mesos_cli/compose/main.py
@@ -77,7 +77,7 @@ class mesosCompose(PluginBase):
             "arguments": ["<framework-name>", "<operations>"],
             "flags": {},
             "short_help": "Framework Commands.",
-            "long_help": "Framework Commands\n\treregister - force the reregistration of the framework. !!! ONLY USE IT DURING MESOS CONNECTION ERRORS. !!!",
+            "long_help": "Framework Commands\n\treregister - force the reregistration of the framework. !!! ONLY USE IT DURING MESOS CONNECTION ERRORS. !!!\n\tsupress - supress the framework\n",
         }
     }
 
@@ -106,6 +106,20 @@ class mesosCompose(PluginBase):
             data = self.write_endpoint(
                 framework_address,
                 "/api/compose/v0/framework/reregister",
+                self,
+                "PUT"
+            )
+            print(data)
+
+        if argv["<operations>"] == "supress":
+            print("Supress the Framework")
+
+            framework_address = get_framework_address(
+                self.get_framework_id(argv), master, config
+            )
+            data = self.write_endpoint(
+                framework_address,
+                "/api/compose/v0/framework/supress",
                 self,
                 "PUT"
             )